+++ ax3l/constants/DReportMgr.py
@@ -2,6 +2,7 @@ from typing import Final
 
 
 class DReportMgr:
+    REFRESH_SECONDS: Final[int] = 30
     PORT: Final[int] = 28870
     PORT_DEV: Final[int] = 28868
     PORT_QA: Final[int] = 28869